[A] Added Create Asset Menu option
Now the Reference To Scene Scriptable Object can be created by right-clicking a Scene asset and selecting Create->Scriptable Objects->Scene Reference.
The option will be enabled if the right-clicked object is a Scene asset and if there is no Reference To Scene object referencing the same scene (compared by GUID).